Melhoria na pasta de testes e novos casos de teste do file transform
